Implement a selinux_port type and provider via semanage
This initial implementation seems to work as expected. The provider
will fail if trying to remove entries defined in the system policy.

The prefetch method feels like a hack, but I am not sure what the best
way to map unique port/protocol pairs to catalog resources is. It must
be done in order to use modify instead of add when there is an existing
context definition.