Load system policy contexts
This change prefetches the system policy contexts. This means that if
the system policy already exists for /var/spool/squid(/.*)? then it
won't attempt to override it when the specification matches. It does
still allow modifications if for some reason the base policy is
incorrect. Purging of all local contexts is also still possible.